# cmd/docs/swagger.yaml
basePath: /api
definitions:
  models.Star:
    properties:
      dec:
        type: number
      description:
        type: string
      imageURL:
        type: string
      isActive:
        type: boolean
      observations:
        description: связь многие-ко-многим через telescope_observation_stars
        items:
          $ref: '#/definitions/models.TelescopeObservation'
        type: array
      ra:
        type: number
      shortDescription:
        type: string
      starID:
        type: integer
      starName:
        type: string
    type: object
  models.TelescopeObservation:
    properties:
      completionDate:
        type: string
      createdAt:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creatorID:
        type: integer
      formationDate:
        type: string
      moderator:
        $ref: '#/definitions/models.User'
      moderatorID:
        type: integer
      observationDate:
        type: string
      observerLatitude:
        type: number
      observerLongitude:
        type: number
      stars:
        items:
          $ref: '#/definitions/models.Star'
        type: array
      status:
        type: string
      telescopeObservationID:
        type: integer
      telescopeObservationStars:
        items:
          $ref: '#/definitions/models.TelescopeObservationStar'
        type: array
    type: object
  models.TelescopeObservationStar:
    properties:
      orderNumber:
        type: integer
      quantity:
        type: integer
      resultValue:
        type: number
      star:
        $ref: '#/definitions/models.Star'
      starID:
        type: integer
      telescopeObservation:
        $ref: '#/definitions/models.TelescopeObservation'
      telescopeObservationID:
        type: integer
    type: object
  models.User:
    properties:
      isModerator:
        type: boolean
      passwordHash:
        type: string
      userID:
        type: integer
      username:
        type: string
    type: object
host: 192.168.1.51:9005
info:
  contact: {}
  description: "Система позволяет пользователям создавать заявки на получение данных
    для расчёта наведения телескопа,\nа модераторам — управлять этими заявками и добавлять
    новые звёзды.\n## \U0001F510 Аутентификация\n- Используются **сессии и cookie**,
    хранящиеся в **Redis**.\n- Без авторизации доступны только методы **чтения (GET)**.\n-
    Для авторизованных запросов браузер автоматически отправляет cookie.\n## \U0001F465
    Роли и права доступа\n- **Гость:** только GET-запросы.\n- **Пользователь:** управление
    своими заявками + чтение данных.\n- **Модератор:** полный доступ ко всем ресурсам."
  title: Calculator Observations Stars API
  version: "1.0"
paths:
  /stars:
    get:
      consumes:
      - application/json
      description: Возвращает список звёзд, доступных для наблюдения. Можно указать
        фильтр по названию.
      parameters:
      - description: Фильтр по названию звезды
        in: query
        name: star_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Star'
            type: array
        "500":
          description: Ошибка получения звёзд
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список звёзд
      tags:
      - Stars
    post:
      consumes:
      - application/json
      description: Создаёт новую запись звезды в базе данных
      parameters:
      - description: Данные звезды
        in: body
        name: star
        required: true
        schema:
          $ref: '#/definitions/models.Star'
      produces:
      - application/json
      responses:
        "201":
          description: Звезда успешно добавлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный JSON или отсутствует название
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при сохранении
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить новую звезду
      tags:
      - Stars
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Логин пользователя и получение токена авторизации
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          properties:
            Password:
              type: string
            Username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid json
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация пользователя
      tags:
      - Users
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Возвращает ID, имя пользователя и флаг модератора для авторизованного
        пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Информация о пользователе
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка получения пользователя
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить информацию о текущем пользователе
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Позволяет обновить имя пользователя или пароль. Флаги модератора
        и ID недоступны для изменения
      parameters:
      - description: Поля для обновления пользователя
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: user updated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid json
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка обновления пользователя
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновление информации о текущем пользователе
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с логином, паролем и флагом модератора
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          properties:
            IsModerator:
              type: boolean
            Password:
              type: string
            Username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: user created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid json
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка создания пользователя
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - Users
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_token
    type: apiKey
swagger: "2.0"